refactor(changelog): move extractor to src module; drop .mjs test import
The Windows CI test job failed deterministically once a test imported a .mjs
script (../scripts/extract-changelog.mjs) — that pattern resolves inconsistently
under vitest/Vite on Windows, while the suite's other tests import source as
../src/...js and pass on all OSes.

- Move the pure logic to src/utils/changelog.ts and import it from the test the
  standard way (../src/utils/changelog.js).
- Replace the .mjs/.d.mts with a thin scripts/extract-changelog.ts CLI run via
  tsx (matching benchmark.ts/validate-import.ts); release.yml invokes it with
  pnpm exec tsx.
- Update the development.md CI guidance to note the src-module + .ts-wrapper
  pattern and to avoid importing .mjs from tests on Windows.
